+++ src/main/java/com/newstoss/news/application/impl/GetRelatedNews.java
@@ -1,8 +1,8 @@
-package com.newstoss.news.application.service;
+package com.newstoss.news.application.impl;
 
 import com.newstoss.news.adapter.in.web.dto.common.RelatedNewsDTO;
 import com.newstoss.news.adapter.out.dto.MLRelatedNewsDTO;
-import com.newstoss.news.application.port.in.GetRelatedNewsUseCase;
+import com.newstoss.news.application.port.in.ml.GetRelatedNewsUseCase;
 import com.newstoss.news.application.port.out.MLNewsRelatedPortV1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
