+++ src/main/java/com/newstoss/news/application/impl/v1/GetRealTimeNewsV1.java
@@ -1,8 +1,8 @@
-package com.newstoss.news.application.service.v1;
+package com.newstoss.news.application.impl.v1;
 
 import com.newstoss.news.adapter.in.web.dto.v1.NewsDTOv1;
 import com.newstoss.news.adapter.out.dto.v1.MLNewsDTOv1;
-import com.newstoss.news.application.port.in.v1.GetRealTimeNewsUseCaseV1;
+import com.newstoss.news.application.port.in.ml.v1.GetRealTimeNewsUseCaseV1;
 import com.newstoss.news.application.port.out.MLNewsRelatedPortV1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
