+++ src/main/java/com/newstoss/news/application/port/in/ml/v2/GetRealTimeNewsUseCaseV2.java
@@ -1,4 +1,4 @@
-package com.newstoss.news.application.port.in.v2;
+package com.newstoss.news.application.port.in.ml.v2;
 
 
 import com.newstoss.news.adapter.in.web.dto.v2.NewsDTOv2;
